pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whatsnew"
version = "0.1.0"
description = "Generate concise changelogs from git history."
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "WhatsNew Maintainers"}]
requires-python = ">=3.10"
dependencies = [
    "requests",
    "PyYAML",
    "rich",
    "tenacity",
    "gitpython",
    "python-dateutil",
]

[project.urls]
Homepage = "https://github.com/whatsnew-cli/whatsnew"

[project.scripts]
whatsnew = "whatsnew.cli:main"
